Add binary, hex, and EEPROM files for TestSketch with bootloader
- Added TestSketch.ino.with_bootloader.bin containing the compiled binary with bootloader.
- Added TestSketch.ino.with_bootloader.hex containing the hex representation of the compiled sketch.
- Added ArduinoUno.bin for EEPROM data storage.

backup

+++ RobotWin/Assets/Scripts/UI/RobotEditor/RobotAnalyzer.cs
@@ -209,7 +209,7 @@ namespace RobotTwin.UI.RobotEditor
                 detail.LoadSafetyFactor = joint.MaxLoadCapacity / appliedLoad;
 
                 // Wear estimation (based on cycles and load)
-                detail.EstimatedCycles = joint.EstimatedLifeCycles;
+                detail.EstimatedCycles = (int)joint.EstimatedLifeCycles;
                 detail.WearFactor = CalculateWearFactor(joint, appliedLoad);
                 detail.MaintenanceInterval = EstimateMaintenanceInterval(joint, detail.WearFactor);
